gh-2386 refactoring
